@atifaziz 2024-05-06
Can I help with any hesitation to merge this? Seems rather straightforward unless I'm missing some JIT intrinsics/tricks that this is relying on and which are not obvious from the lack of IL changes.