@lahma 2024-05-07
I think the general rule is that if the struct is more than int pointer size it *might* make a difference. On x64 this struct is borderline so if no fields are anticipated to be added, might not need the strict requirement of being passed as `ref`.

Not sure what the current restrictions are though?