@sebastienros 2024-05-07
I ran the benchmarks and the changes are small, but much more on a single Parser call (SkipWhiteSpace1/10):

Before

```
| Method                         | Mean         | Error          | StdDev       | Gen0    | Gen1   | Allocated |
|------------------------------- |-------------:|---------------:|-------------:|--------:|-------:|----------:|
| CreateCompiledSmallParser      | 38,663.27 ns | 111,132.125 ns | 6,091.527 ns |  3.4180 | 3.1738 |   33289 B |
| CreateCompiledExpressionParser |  6,117.24 ns |   1,458.822 ns |    79.963 ns |  0.5188 | 0.4883 |    5024 B |
|                                |              |                |              |         |        |           |
| CursorMatchHello               |     20.78 ns |       8.627 ns |     0.473 ns |  0.0136 |      - |     128 B |
| CursorMatchGoodbye             |     20.18 ns |       1.549 ns |     0.085 ns |  0.0136 |      - |     128 B |
| CursorMatchNone                |     18.66 ns |      22.188 ns |     1.216 ns |  0.0136 |      - |     128 B |
|                                |              |                |              |         |        |           |
| DecodeStringWithoutEscapes     |     11.09 ns |       0.640 ns |     0.035 ns |       - |      - |         - |
| DecodeStringWithEscapes        |     38.97 ns |       2.381 ns |     0.131 ns |  0.0127 | 0.0001 |     120 B |
|                                |              |                |              |         |        |           |
| ExpressionRawBig               |    839.66 ns |     140.574 ns |     7.705 ns |  0.1268 |      - |    1200 B |
| ExpressionCompiledBig          |  1,502.99 ns |     446.331 ns |    24.465 ns |  0.3052 |      - |    2888 B |
| ExpressionFluentBig            |  1,901.86 ns |     818.213 ns |    44.849 ns |  0.3052 |      - |    2888 B |
|                                |              |                |              |         |        |           |
| ExpressionRawSmall             |    164.20 ns |      44.671 ns |     2.449 ns |  0.0322 |      - |     304 B |
| ExpressionCompiledSmall        |    290.48 ns |     124.043 ns |     6.799 ns |  0.0696 |      - |     656 B |
| ExpressionFluentSmall          |    364.68 ns |     115.644 ns |     6.339 ns |  0.0696 |      - |     656 B |
|                                |              |                |              |         |        |           |
| BigJson                        | 51,068.12 ns |  23,695.190 ns | 1,298.813 ns |  9.9487 | 1.3428 |   93992 B |
| BigJsonCompiled                | 47,148.08 ns |   2,966.961 ns |   162.629 ns |  9.9487 | 1.7700 |   93992 B |
|                                |              |                |              |         |        |           |
| DeepJson                       | 34,704.36 ns |   7,603.367 ns |   416.766 ns | 10.6812 | 1.1597 |  100688 B |
| DeepJsonCompiled               | 28,255.66 ns |   3,254.650 ns |   178.398 ns | 10.6812 | 2.0142 |  100688 B |
|                                |              |                |              |         |        |           |
| LongJson                       | 45,188.37 ns |   7,572.066 ns |   415.050 ns | 13.0615 | 2.6245 |  123224 B |
| LongJsonCompiled               | 39,278.41 ns |  10,857.236 ns |   595.122 ns | 13.0615 | 2.6245 |  123224 B |
|                                |              |                |              |         |        |           |
| WideJson                       | 21,991.07 ns |   5,252.902 ns |   287.929 ns |  4.3945 | 0.3967 |   41536 B |
| WideJsonCompiled               | 20,914.31 ns |   6,822.441 ns |   373.961 ns |  4.3945 | 0.4883 |   41536 B |
|                                |              |                |              |         |        |           |
| Lookup                         |     22.37 ns |       4.576 ns |     0.251 ns |  0.0136 |      - |     128 B |
|                                |              |                |              |         |        |           |
| SkipWhiteSpace_1               |     15.85 ns |       2.030 ns |     0.111 ns |  0.0136 |      - |     128 B |
| SkipWhiteSpace_10              |     21.37 ns |       7.068 ns |     0.387 ns |  0.0136 |      - |     128 B |
```

After

```
| Method                         | Mean         | Error         | StdDev       | Gen0    | Gen1   | Allocated |
|------------------------------- |-------------:|--------------:|-------------:|--------:|-------:|----------:|
| CreateCompiledSmallParser      | 33,481.15 ns | 34,989.242 ns | 1,917.878 ns |  3.4180 | 3.1738 |   33289 B |
| CreateCompiledExpressionParser |  6,276.06 ns |  6,440.387 ns |   353.019 ns |  0.5188 | 0.4883 |    5024 B |
|                                |              |               |              |         |        |           |
| CursorMatchHello               |     20.23 ns |      0.479 ns |     0.026 ns |  0.0136 |      - |     128 B |
| CursorMatchGoodbye             |     23.51 ns |     46.454 ns |     2.546 ns |  0.0136 |      - |     128 B |
| CursorMatchNone                |     17.33 ns |      1.315 ns |     0.072 ns |  0.0136 |      - |     128 B |
|                                |              |               |              |         |        |           |
| DecodeStringWithoutEscapes     |     11.14 ns |      1.654 ns |     0.091 ns |       - |      - |         - |
| DecodeStringWithEscapes        |     40.39 ns |     31.777 ns |     1.742 ns |  0.0127 | 0.0001 |     120 B |
|                                |              |               |              |         |        |           |
| ExpressionRawBig               |    830.66 ns |     66.513 ns |     3.646 ns |  0.1268 |      - |    1200 B |
| ExpressionCompiledBig          |  1,548.18 ns |    447.288 ns |    24.517 ns |  0.3052 |      - |    2888 B |
| ExpressionFluentBig            |  2,021.56 ns |    424.065 ns |    23.244 ns |  0.3052 |      - |    2888 B |
|                                |              |               |              |         |        |           |
| ExpressionRawSmall             |    168.86 ns |     25.437 ns |     1.394 ns |  0.0322 |      - |     304 B |
| ExpressionCompiledSmall        |    295.55 ns |     62.882 ns |     3.447 ns |  0.0696 |      - |     656 B |
| ExpressionFluentSmall          |    360.47 ns |     69.248 ns |     3.796 ns |  0.0696 |      - |     656 B |
|                                |              |               |              |         |        |           |
| BigJson                        | 51,121.76 ns |  5,311.321 ns |   291.131 ns |  9.9487 | 1.3428 |   93992 B |
| BigJsonCompiled                | 48,300.99 ns |  9,526.195 ns |   522.163 ns |  9.9487 | 1.7700 |   93992 B |
|                                |              |               |              |         |        |           |
| DeepJson                       | 33,876.91 ns |    806.898 ns |    44.229 ns | 10.6812 | 1.1597 |  100688 B |
| DeepJsonCompiled               | 28,366.68 ns |  1,299.791 ns |    71.246 ns | 10.6812 | 2.0142 |  100688 B |
|                                |              |               |              |         |        |           |
| LongJson                       | 43,190.38 ns |  4,418.623 ns |   242.200 ns | 13.0615 | 2.6245 |  123224 B |
| LongJsonCompiled               | 40,101.58 ns | 13,241.494 ns |   725.811 ns | 13.0615 | 2.6245 |  123224 B |
|                                |              |               |              |         |        |           |
| WideJson                       | 22,854.24 ns | 10,767.236 ns |   590.189 ns |  4.3945 | 0.3967 |   41536 B |
| WideJsonCompiled               | 22,378.38 ns |  5,584.930 ns |   306.129 ns |  4.3945 | 0.4883 |   41536 B |
|                                |              |               |              |         |        |           |
| Lookup                         |     29.76 ns |     32.111 ns |     1.760 ns |  0.0136 |      - |     128 B |
|                                |              |               |              |         |        |           |
| SkipWhiteSpace_1               |     24.28 ns |     39.807 ns |     2.182 ns |  0.0136 |      - |     128 B |
| SkipWhiteSpace_10              |     27.76 ns |     39.446 ns |     2.162 ns |  0.0136 |      - |     128 B |
```